+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Extracted icon renderer to Go-based binary for major performance improvement.
+- Improved error message in the case that Factorio crashed.
 
 ### Fixed